Fix warnings in Sav3

+++ source/sav/Sav3.cpp
@@ -60,9 +60,9 @@ int Sav3::getActiveSaveIndex(std::shared_ptr<u8[]> dt, std::array<int, BLOCK_COU
 {
     int zeroBlock1 = std::find(blockOrder1.begin(), blockOrder1.end(), 0) - blockOrder1.begin();
     int zeroBlock2 = std::find(blockOrder2.begin(), blockOrder2.end(), 0) - blockOrder2.begin();
-    if (zeroBlock2 == blockOrder2.size())
+    if ((size_t)zeroBlock2 == blockOrder2.size())
         return 0;
-    if (zeroBlock1 == blockOrder1.size())
+    if ((size_t)zeroBlock1 == blockOrder1.size())
         return 1;
     u32 count1 = LittleEndian::convertTo<u32>(&dt[(zeroBlock1 * SIZE_BLOCK) + 0x0FFC]);
     u32 count2 = LittleEndian::convertTo<u32>(&dt[(zeroBlock2 * SIZE_BLOCK) + 0xEFFC]);
